Update README.md
just converting some text to links, it helps with getting to the InfiniTime GitHub page and firmware releases pages quicker after clicking the Infini-iOS GitHub page link in the actual app.  May help others as a stop gap until a decision is made on having online firmware updates is made.