Update SettingsView.swift
I have no clue if this is all that is needed to add the link on the Settings section of the app, apologies if it is not.  :(

Figured adding a link to the firmware releases pages may help others as a way to quickly get to and download newer firmware to test until a decision is made on having online firmware updates is made.